fix: Correción de typo de datos en DTO de roles

+++ ed-lib-common/src/main/java/com/eduplanner/ed_lib_common/dto/RegisterStudentDTO.java
@@ -35,7 +35,7 @@ public class RegisterStudentDTO {
 
     private String disabilities;
 
-    private Byte stratum;
+    private Integer stratum;
 
     private String populationType;
